LayoutTests/http/tests/loading/preload-picture-type-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
main frame - willPerformClientRedirectToURL: http://127.0.0.1:8000/loading/preload-picture-type.html 
main frame - didFinishDocumentLoadForFrame
main frame - didFinishLoadForFrame
main frame - didStartProvisionalLoadForFrame
main frame - didCancelClientRedirectForFrame
main frame - didCommitLoadForFrame
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
PASS window.devicePixelRatio is 1
PASS internals.isPreloaded('resources/base-image1.png?0'); is true
PASS internals.isPreloaded('resources/preload-test.jpg?0'); is false
PASS internals.isPreloaded('resources/base-image1.png?1'); is true
PASS internals.isPreloaded('resources/preload-test.jpg?1'); is false
PASS internals.isPreloaded('resources/base-image1.png?2'); is true
PASS internals.isPreloaded('resources/preload-test.jpg?2'); is false
PASS internals.isPreloaded('resources/base-image1.png?3'); is false
PASS internals.isPreloaded('resources/preload-test.jpg?3'); is true
PASS internals.isPreloaded('resources/base-image1.png?4'); is false
PASS internals.isPreloaded('resources/base-image2.png?4'); is false
PASS internals.isPreloaded('resources/base-image3.png?4'); is true
PASS internals.isPreloaded('resources/preload-test.jpg?4'); is false
PASS internals.isPreloaded('resources/base-image1.png?5'); is false
PASS internals.isPreloaded('resources/base-image2.png?5'); is false
PASS internals.isPreloaded('resources/base-image3.png?5'); is false
PASS internals.isPreloaded('resources/preload-test.jpg?5'); is true
PASS internals.isPreloaded('resources/base-image1.png?6'); is true
PASS internals.isPreloaded('resources/preload-test.jpg?6'); is false
PASS internals.isPreloaded('resources/base-image1.png?7'); is false
PASS internals.isPreloaded('resources/preload-test.jpg?7'); is true
PASS internals.isPreloaded('resources/base-image1.png?8'); is true
PASS internals.isPreloaded('resources/preload-test.jpg?8'); is false
PASS internals.isPreloaded('resources/base-image1.png?9'); is false
PASS internals.isPreloaded('resources/preload-test.jpg?9'); is true
PASS internals.isPreloaded('resources/base-image1.png?10'); is true
PASS internals.isPreloaded('resources/base-image2.png?10'); is false
PASS internals.isPreloaded('resources/preload-test.jpg?10'); is false
PASS internals.isPreloaded('resources/base-image1.png?11'); is false
PASS internals.isPreloaded('resources/base-image2.png?11'); is true
PASS internals.isPreloaded('resources/preload-test.jpg?11'); is false
PASS internals.isPreloaded('resources/base-image1.png?12'); is true
PASS internals.isPreloaded('resources/base-image2.png?12'); is false
PASS internals.isPreloaded('resources/preload-test.jpg?12'); is false
PASS internals.isPreloaded('resources/base-image1.png?13'); is false
PASS internals.isPreloaded('resources/base-image2.png?13'); is false
PASS internals.isPreloaded('resources/preload-test.jpg?13'); is true
PASS successfullyParsed is true

TEST COMPLETE
                   
